+++ csobpg/v19/models/actions.py
@@ -31,10 +31,7 @@ class Endpoint(_s.SignedModel):
         return (
             self.url,
             self.method,
-            # NOTE: it is not clear how to handle the `vars` parameter in the
-            # signature. We assume that the signature is generated from the
-            # values
-            *(self.vars.values() if self.vars else ()),
+            self.vars,
         )
 
     def __str__(self) -> str: